Moved some files from lib to utilities // Started adding comments // Was working on testing the emailing functionality and became blocked by bug with JSON schema
