--- wishlist-template/api/lib/index.js
@@ -1,1 +1,0 @@
-export { default as updateWishlistMetafield } from "./updateWishlistMetafield";
