+++ wishlist-template/api/models/wishlistItem/actions/delete.js
@@ -3,7 +3,7 @@ import {
   ActionOptions,
   DeleteWishlistItemActionContext,
 } from "gadget-server";
-import { updateWishlistMetafield } from "../../../lib";
+import { updateWishlistMetafield } from "../../../utilities";
 
 /**
  * @param { DeleteWishlistItemActionContext } context
