+++ wishlist-template/api/utilities/index.js
@@ -1,1 +1,2 @@
 export { Email } from "./email";
+export { default as updateWishlistMetafield } from "./updateWishlistMetafield";
